\name{SCS}
\alias{SCS}

\docType{data}

\title{Sexual Compulsivity Scale Data Set}

\description{
The items were rated on a likert scale (1=Not at all like me, 2=Slightly like me, 3=Mainly like me, 4=Very much like me):
\itemize{
\item \eqn{Q1}: My sexual appetite has gotten in the way of my relationships.			
\item \eqn{Q2}: My sexual thoughts and behaviors are causing problems in my life.			
\item \eqn{Q3}: My desires to have sex have disrupted my daily life.				
\item \eqn{Q4}: I sometimes fail to meet my commitments and responsibilities because of my sexual behaviors.		
\item \eqn{Q5}: I sometimes get so horny I could lose control.				
\item \eqn{Q6}: I find myself thinking about sex while at work.				
\item \eqn{Q7}: I feel that sexual thoughts and feelings are stronger than I am.			
\item \eqn{Q8}: I have to struggle to control my sexual thoughts and behavior.			
\item \eqn{Q9}: I think about sex more than I would like to.				
\item \eqn{Q10}. It has been difficult for me to find sex partners who desire having sex as much as I want to.
\item \eqn{Gender}: chosen from drop down list (1=male, 2=female).
}
}

\format{
 The \code{SCS} matrix consists of 3215 rows (one per subject) and 11 columns (one per item). 
}

\source{
The full dataset is available at the following URL: \emph{https://openpsychometrics.org/_rawdata/} }

\references{
Kalichman, S. C., & Rompa, D. (1995). Sexual sensation seeking and sexual compulsivity scales: Reliability, validity, and predicting HIV risk behavior. \emph{Journal of Personality Assessment}, 65(3), 586–601. https://doi.org/10.1207/s15327752jpa6503_16
}


